+++ Runtime/IPurchaseResponse.cs
@@ -5,6 +5,7 @@ namespace DTech.InAppFlex
     public interface IPurchaseResponse
     {
         Product Product { get; }
+        string ProductId { get; }
         string TransactionId { get; }
         string Receipt { get; }
         PurchaseStatus Status { get; }